pyproject.toml
[tool.poetry]
name = "paper-orbit-scraper"
version = "0.1.0"
description = "FastAPI project with Poetry"
authors = ["Developer"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.116.1"
uvicorn = {extras = ["standard"], version = "^0.35.0"}
playwright = "^1.48.0"
pydantic = "^2.9.0"
python-dotenv = "^1.1.1"
cryptography = "^45.0.6"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
